move DatePicker into separate component file
